requirements.txt
pip
pillow
imageio
exifread
PyYAML
